+++ src/oci/psql/models/backup_source_details.py
@@ -23,7 +23,7 @@ class BackupSourceDetails(SourceDetails):
 
         :param source_type:
             The value to assign to the source_type property of this BackupSourceDetails.
-            Allowed values for this property are: "BACKUP", "NONE", "DB_SYSTEM"
+            Allowed values for this property are: "BACKUP", "NONE", "POINT_IN_TIME_DB_SYSTEM", "DB_SYSTEM"
         :type source_type: str
 
         :param backup_id: