+++ src/oci/resource_manager/models/bitbucket_cloud_email_api_token_configuration_source_provider.py
@@ -10,74 +10,82 @@ from oci.decorators import init_model_state_from_kwargs
 
 
 @init_model_state_from_kwargs
-class BitbucketCloudUsernameAppPasswordConfigurationSourceProvider(ConfigurationSourceProvider):
+class BitbucketCloudEmailApiTokenConfigurationSourceProvider(ConfigurationSourceProvider):
     """
-    The properties that define a configuration source provider of the type `BITBUCKET_CLOUD_USERNAME_APPPASSWORD`.
-    This type corresponds to a configuration source provider in Bitbucket cloud that is authenticated with a username and app password.
+    The properties that define a configuration source provider of the type `BITBUCKET_CLOUD_ACCESS_TOKEN`.
+    This type corresponds to a configuration source provider in Bitbucket Cloud that is authenticated with Atlassian account email and API token.
     """
 
     def __init__(self, **kwargs):
         """
-        Initializes a new BitbucketCloudUsernameAppPasswordConfigurationSourceProvider object with values from keyword arguments. The default value of the :py:attr:`~oci.resource_manager.models.BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.config_source_provider_type` attribute
-        of this class is ``BITBUCKET_CLOUD_USERNAME_APPPASSWORD`` and it should not be changed.
+        Initializes a new BitbucketCloudEmailApiTokenConfigurationSourceProvider object with values from keyword arguments. The default value of the :py:attr:`~oci.resource_manager.models.BitbucketCloudEmailApiTokenConfigurationSourceProvider.config_source_provider_type` attribute
+        of this class is ``BITBUCKET_CLOUD_ACCESS_TOKEN`` and it should not be changed.
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
         :param id:
-            The value to assign to the id property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the id property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type id: str
 
         :param compartment_id:
-            The value to assign to the compartment_id property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the compartment_id property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type compartment_id: str
 
         :param display_name:
-            The value to assign to the display_name property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the display_name property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type display_name: str
 
         :param description:
-            The value to assign to the description property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the description property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type description: str
 
         :param time_created:
-            The value to assign to the time_created property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the time_created property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type time_created: datetime
 
         :param lifecycle_state:
-            The value to assign to the lifecycle_state property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the lifecycle_state property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
             Allowed values for this property are: "ACTIVE"
         :type lifecycle_state: str
 
         :param config_source_provider_type:
-            The value to assign to the config_source_provider_type property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
-            Allowed values for this property are: "BITBUCKET_CLOUD_USERNAME_APPPASSWORD", "BITBUCKET_SERVER_ACCESS_TOKEN", "GITLAB_ACCESS_TOKEN", "GITHUB_ACCESS_TOKEN"
+            The value to assign to the config_source_provider_type property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
+            Allowed values for this property are: "BITBUCKET_CLOUD_ACCESS_TOKEN", "BITBUCKET_SERVER_ACCESS_TOKEN", "GITLAB_ACCESS_TOKEN", "GITHUB_ACCESS_TOKEN"
         :type config_source_provider_type: str
 
         :param private_server_config_details:
-            The value to assign to the private_server_config_details property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the private_server_config_details property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type private_server_config_details: oci.resource_manager.models.PrivateServerConfigDetails
 
         :param username:
-            The value to assign to the username property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the username property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type username: str
 
+        :param email:
+            The value to assign to the email property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
+        :type email: str
+
+        :param is_migration_required:
+            The value to assign to the is_migration_required property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
+        :type is_migration_required: bool
+
         :param secret_id:
-            The value to assign to the secret_id property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the secret_id property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type secret_id: str
 
         :param freeform_tags:
-            The value to assign to the freeform_tags property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the freeform_tags property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type freeform_tags: dict(str, str)
 
         :param defined_tags:
-            The value to assign to the defined_tags property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the defined_tags property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type defined_tags: dict(str, dict(str, object))
 
         :param system_tags:
-            The value to assign to the system_tags property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the system_tags property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type system_tags: dict(str, dict(str, object))
 
         :param api_endpoint:
-            The value to assign to the api_endpoint property of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+            The value to assign to the api_endpoint property of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type api_endpoint: str
 
         """
@@ -91,6 +99,8 @@ class BitbucketCloudUsernameAppPasswordConfigurationSourceProvider(Configuration
             'config_source_provider_type': 'str',
             'private_server_config_details': 'PrivateServerConfigDetails',
             'username': 'str',
+            'email': 'str',
+            'is_migration_required': 'bool',
             'secret_id': 'str',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))',
@@ -107,6 +117,8 @@ class BitbucketCloudUsernameAppPasswordConfigurationSourceProvider(Configuration
             'config_source_provider_type': 'configSourceProviderType',
             'private_server_config_details': 'privateServerConfigDetails',
             'username': 'username',
+            'email': 'email',
+            'is_migration_required': 'isMigrationRequired',
             'secret_id': 'secretId',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags',
@@ -122,22 +134,24 @@ class BitbucketCloudUsernameAppPasswordConfigurationSourceProvider(Configuration
         self._config_source_provider_type = None
         self._private_server_config_details = None
         self._username = None
+        self._email = None
+        self._is_migration_required = None
         self._secret_id = None
         self._freeform_tags = None
         self._defined_tags = None
         self._system_tags = None
         self._api_endpoint = None
-        self._config_source_provider_type = 'BITBUCKET_CLOUD_USERNAME_APPPASSWORD'
+        self._config_source_provider_type = 'BITBUCKET_CLOUD_ACCESS_TOKEN'
 
     @property
     def api_endpoint(self):
         """
-        Gets the api_endpoint of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+        Gets the api_endpoint of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         The Bitbucket cloud service endpoint.
         Example: `https://bitbucket.org/`
 
 
-        :return: The api_endpoint of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+        :return: The api_endpoint of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :rtype: str
         """
         return self._api_endpoint
@@ -145,12 +159,12 @@ class BitbucketCloudUsernameAppPasswordConfigurationSourceProvider(Configuration
     @api_endpoint.setter
     def api_endpoint(self, api_endpoint):
         """
-        Sets the api_endpoint of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+        Sets the api_endpoint of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         The Bitbucket cloud service endpoint.
         Example: `https://bitbucket.org/`
 
 
-        :param api_endpoint: The api_endpoint of this BitbucketCloudUsernameAppPasswordConfigurationSourceProvider.
+        :param api_endpoint: The api_endpoint of this BitbucketCloudEmailApiTokenConfigurationSourceProvider.
         :type: str
         """
         self._api_endpoint = api_endpoint