+++ src/oci/resource_manager/models/update_configuration_source_provider_details.py
@@ -20,9 +20,9 @@ class UpdateConfigurationSourceProviderDetails(object):
         Initializes a new UpdateConfigurationSourceProviderDetails object with values from keyword arguments. This class has the following subclasses and if you are using this class as input
         to a service operations then you should favor using a subclass over the base class:
 
-        * :class:`~oci.resource_manager.models.UpdateBitbucketCloudUsernameAppPasswordConfigurationSourceProviderDetails`
         * :class:`~oci.resource_manager.models.UpdateBitbucketServerAccessTokenConfigurationSourceProviderDetails`
         * :class:`~oci.resource_manager.models.UpdateGitlabAccessTokenConfigurationSourceProviderDetails`
+        * :class:`~oci.resource_manager.models.UpdateBitbucketCloudEmailApiTokenConfigurationSourceProviderDetails`
         * :class:`~oci.resource_manager.models.UpdateGithubAccessTokenConfigurationSourceProviderDetails`
 
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
@@ -83,14 +83,14 @@ class UpdateConfigurationSourceProviderDetails(object):
         """
         type = object_dictionary['configSourceProviderType']
 
-        if type == 'BITBUCKET_CLOUD_USERNAME_APPPASSWORD':
-            return 'UpdateBitbucketCloudUsernameAppPasswordConfigurationSourceProviderDetails'
-
         if type == 'BITBUCKET_SERVER_ACCESS_TOKEN':
             return 'UpdateBitbucketServerAccessTokenConfigurationSourceProviderDetails'
 
         if type == 'GITLAB_ACCESS_TOKEN':
             return 'UpdateGitlabAccessTokenConfigurationSourceProviderDetails'
+
+        if type == 'BITBUCKET_CLOUD_ACCESS_TOKEN':
+            return 'UpdateBitbucketCloudEmailApiTokenConfigurationSourceProviderDetails'
 
         if type == 'GITHUB_ACCESS_TOKEN':
             return 'UpdateGithubAccessTokenConfigurationSourceProviderDetails'
@@ -150,7 +150,8 @@ class UpdateConfigurationSourceProviderDetails(object):
         """
         Gets the config_source_provider_type of this UpdateConfigurationSourceProviderDetails.
         The type of configuration source provider.
-        The `BITBUCKET_CLOUD_USERNAME_APPPASSWORD` type corresponds to Bitbucket Cloud.
+        The `BITBUCKET_CLOUD_ACCESS_TOKEN` type corresponds to Bitbucket Cloud.
+        For Bitbucket Cloud, update requests must provide Atlassian account `email` and `secretId` containing an API token.
         The `BITBUCKET_SERVER_ACCESS_TOKEN` type corresponds to Bitbucket Server.
         The `GITLAB_ACCESS_TOKEN` type corresponds to GitLab.
         The `GITHUB_ACCESS_TOKEN` type corresponds to GitHub.
@@ -166,7 +167,8 @@ class UpdateConfigurationSourceProviderDetails(object):
         """
         Sets the config_source_provider_type of this UpdateConfigurationSourceProviderDetails.
         The type of configuration source provider.
-        The `BITBUCKET_CLOUD_USERNAME_APPPASSWORD` type corresponds to Bitbucket Cloud.
+        The `BITBUCKET_CLOUD_ACCESS_TOKEN` type corresponds to Bitbucket Cloud.
+        For Bitbucket Cloud, update requests must provide Atlassian account `email` and `secretId` containing an API token.
         The `BITBUCKET_SERVER_ACCESS_TOKEN` type corresponds to Bitbucket Server.
         The `GITLAB_ACCESS_TOKEN` type corresponds to GitLab.
         The `GITHUB_ACCESS_TOKEN` type corresponds to GitHub.