+++ CHANGELOG.rst
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 ====================
+2.181.0 - 2026-06-30
+====================
+
+Added 
+----- 
+* Support for Bitbucket Cloud email and API token configuration source providers in the Resource Manager service 
+* Support for configurable randomization windows for schedules in the Data Science service 
+* Support for partner-provided Windows licensing in the Compute service 
+* Support for Key Management service (KMS) bucket key in the Object Storage service 
+* Support for Point-In-Time Recovery (PITR) in the PostgreSQL service  
+
+Breaking 
+-------- 
+* The models `BitbucketCloudUsernameAppPasswordConfigurationSourceProviderSummary`, `BitbucketCloudUsernameAppPasswordConfigurationSourceProvider`,  `CreateBitbucketCloudUsernameAppPasswordConfigurationSourceProviderDetails` and `UpdateBitbucketCloudUsernameAppPasswordConfigurationSourceProviderDetails` were removed in the Resource Manager service  
+
+====================
 2.180.0 - 2026-06-23
 ====================